gemini-vm-frontend/requirements.txt
Flask
google-generativeai
google-cloud-compute
python-dotenv
requests # For making HTTP requests to the sompaak/runner
